move debounce to common/timing
