+++ ui/common/src/resize.ts
@@ -1,6 +1,6 @@
 import * as cg from 'chessground/types';
 import * as xhr from './xhr';
-import debounce from './debounce';
+import { debounce } from './timing';
 import * as Prefs from './prefs';
 
 type MouchEvent = Event & Partial<MouseEvent & TouchEvent>;
