+++ ui/mod/src/mod.user.ts
@@ -1,5 +1,5 @@
 import * as xhr from 'common/xhr';
-import debounce from 'common/debounce';
+import { debounce } from 'common/timing';
 import * as licon from 'common/licon';
 import extendTablesortNumber from 'common/tablesortNumber';
 import tablesort from 'tablesort';
